For #9032 tests(nimbus): Move test into remote settings
